requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.0

openai==1.45.0
tenacity==9.0.0
twilio==9.2.3

pydantic==2.8.2
loguru==0.7.2

SQLAlchemy==2.0.32
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

beautifulsoup4==4.12.3
playwright==1.53.0
python-multipart==0.0.9

reportlab==4.2.2
openpyxl==3.1.5
